// Frontend/src/app/models/welcome-plans.ts
/**************************************************************************

   File:       welcome-plans-json.ts
   Copyright:   (c) 2023 NazImposter

   Module-History:
   Date         Author                Reason
   03.12.2023   Pasa Larisa      Basic structure

 **************************************************************************/

   export interface Plans {
    namePlan: string;
    planForClientType: string;
    cloudStorage: string;
    fitFor: string;
    description: string;
  }